#16: rename pagemodal.dart to photopage.dart
